new: Stop claiming support of option -p ab
The "new" command supports option -p ab, which will add -p ab to the
new patch's options in the series file. Problem is, this option isn't
actually supported in the series file. Only options which should be
passed to the patch command are supported.

So don't let the user pass option -p ab when creating a patch.

This should prevent some confusion as reported here:
https://savannah.nongnu.org/patch/index.php?10094

Fixes: 66f9da46333e ("new command: Add -p ... option (equivalent to diff -p ...)")
Signed-off-by: Jean Delvare <jdelvare@suse.de>