patches: Performance boost on large series
On large series, the patches command is horribly slow. This is caused
by a performance issue in bash 4.4 and older when a function with a
lot of positional parameters calls another function.

This can be worked around by expanding the list of patches inside the
functions instead of passing them as parameters. In my tests, this
speeds up the command by a factor 1000 (from over 34 minutes to 2
seconds on a fully applied 48000 patch series).

Signed-off-by: Jean Delvare <jdelvare@suse.de>