setup: New option --spec-filter
Some build systems (such as OBS) don't work with compliant rpm spec
files. Instead, they use templates, which they preprocess before
passing them to rpmbuild. In order to be able to use "quilt setup"
on these spec files, we need to perform the same kind of
preprocessing.

For this purpose, introduce a new option to the setup command,
--spec-filter. This lets the user provide a customer filtering script
to preprocess their spec file before rpmbuild gets called on it.

Signed-off-by: Jean Delvare <jdelvare@suse.de>