setup: Fix for rpm 4.20's RPM_BUILD_ROOT change
rpm 4.20 creates a working directory (named
${RPM_PACKAGE_NAME}-${RPM_PACKAGE_VERSION}-build) inside _builddir
and RPM_BUILD_DIR points to that directory instead of _builddir as
it used to. This confuses the inspect-wrapper script which assumes
that RPM_BUILD_DIR points to _builddir.

Detect this case and make RPM_BUILD_DIR point to _builddir again so
that the rest of the code keeps working.

Signed-off-by: Jean Delvare <jdelvare@suse.de>